esra-explain-worker/requirements.txt
elastic-transport==8.4.0
easy-elasticsearch==0.0.9
Flask==2.2.2
keras==2.11.0
nlp==0.4.0
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.6
spacy==3.3.1
spacy-legacy==3.0.11
spacy-loggers==1.0.4
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
gevent==22.10.2
nltk==3.8.1
transformers==4.26.0
sentence-transformers==2.2.2
faiss-cpu==1.7.3
faiss-gpu==1.7.2
pdfplumber==0.8.0
beir==1.0.0